added default methods, adjusted implementation in OffensesFilter and EmodgiFilter
