+++ simple-chat/src/main/java/com/solvd/automation/util/SerializationUtil.java
@@ -54,57 +54,3 @@ public class SerializationUtil {
 
 }
 
-//public class SerializationUtil {
-//    private String dataFilePath  = System.getProperty("user.dir") + "\\src\\serial";
-//    private String dataFileResponse = System.getProperty("user.dir") + "\\src\\serial_response";
-//
-//    private static ObjectReader objectReader = new ObjectReader();
-//    private static  ObjectReader objectReaderResponse;
-//
-////    public SerializationUtil(String dataFilePath, String dataFileResponse) {
-////        this.dataFilePath = dataFilePath;
-////        this.dataFileResponse = dataFileResponse;
-////
-////        this.objectReader = new ObjectReader(dataFilePath);
-////        this.objectReaderResponse = new ObjectReader(dataFileResponse);
-////    }
-//
-//    public static void writeObject(Packable obj) {
-//        try {
-//            new ObjectWriter().write(dataFilePath, obj);
-//        } catch (UnableToWriteException e) {
-//            e.printStackTrace();
-//            throw new RuntimeException(String.format("%s is unable to write!", dataFilePath));
-//        }
-//    }
-//
-//    public static Packable readObject() {
-//        try {
-//            return objectReader.read();
-//        } catch (UnableToReadException e) {
-//            e.printStackTrace();
-//            throw new RuntimeException(String.format("%s is unable to read!", dataFilePath));
-//        }
-//
-//    }
-//
-//    public static Packable readResponse() {
-//        try {
-//            return objectReaderResponse.read();
-//        } catch (UnableToReadException e) {
-//            e.printStackTrace();
-//            throw new RuntimeException(String.format("%s is unable to read!", dataFileResponse));
-//        }
-//
-//    }
-//
-//    public static void writeResponse(Packable obj) {
-//        try {
-//            new ObjectWriter().write(dataFileResponse, obj);
-//        } catch (UnableToWriteException e) {
-//            e.printStackTrace();
-//            throw new RuntimeException(String.format("%s is unable to write!", dataFileResponse));
-//        }
-//    }
-//
-//}
