+++ simple-chat/src/main/resources/name.txt
@@ -10,3 +10,4 @@ Elijah
 Ethan
 Noah
 George
+Olga
